Refactor extension path
