+++ app/src/main/java/io/asanre/app/core/data/HttpExtensions.kt
@@ -1,6 +1,6 @@
 package io.asanre.app.core.data
 
-import io.asanre.app.core.domain.Try
+import io.asanre.app.core.domain.extensions.Try
 import io.ktor.client.HttpClient
 import io.ktor.client.call.body
 import io.ktor.client.plugins.resources.get
